Update Day 10.py
Removed a leftover comment left behind that was intended to mark something to come back to later (it was right at the time it was written, but of course a major drawback with comments is that the compiler/interpreter doesn't read them...)

Also some minor formatting updates to other comments.

There's something magical about the submit button that makes it easy to find these sorts of things.