Day 12 python, improved Day 8 C++/D, added timer.h
Day 12 Python is added.

Fixed a probable bug in Day 8 D that would likely result in issues with certain datasets due to an incorrect loop condition that happened to not be an issue with the input I'd given it.

Major refactoring/near-rewrite of Day 8 C++, based largely on what changed structurally overall between writing the C++ solution and the D solution. It should be both faster and easier to read.

Added timer.h to common. Generally, it'll go unused in the final uploads, but is sometimes used during development for benchmarking purposes.