Update Day_18.py
At some point, Day_18.py got broken and I never noticed. Looking through the history it looks like a commit was made editing a comment, but a line of code is also in there that seems like it shouldn't be there at all. No idea how it could have gotten there, as I'm guessing if I didn't notice the breakage I must not have thought it was necessary to test for some reason, so I'm guessing I probably only intended to edit comments.